:recycle: feat: add interfaces
