+++ src/components/ProductList/index.tsx
@@ -1,30 +1,16 @@
-import React, { useEffect, useState } from "react";
-import axios from 'axios';
-import ProductItem from "../productItem";
 import './index.css'
 
-const API = 'https://api.escuelajs.co/api/v1/products';
+import ProductItem from "../productItem";
 
-const ProductList = () => {
+import {ProductListProps} from '../../interfaces';
 
-    const [products, setProducts] = useState([]);
-
-    useEffect(() => {
-        const getProduct = async() => {
-            const response = await axios(API);
-            if(response) {
-                setProducts(response.data)
-            }
-        }
-        getProduct()
-    }, [])
-
+const ProductList = ({products}: ProductListProps) => {
     return (
         <section className="productList__container">
             <div className="productList__list">
                 {products.map(({id, title, price, images}) => (
                         <div className="colum" key={id}>
-                            <ProductItem id={id} price={price} title={title} image={images[0]} />
+                            <ProductItem id={id} price={price} title={title} images={images[0]} />
                         </div>
                     ))}
             </div>
