--- src/components/atomos/Image.tsx
@@ -1,9 +1,0 @@
-const Image = () => {
-    return(
-        <div>
-            Imagen
-        </div>
-    )
-}
-
-export default Image
