+++ src/containers/Layout.tsx
@@ -1,8 +1,4 @@
-import {ReactNode} from 'react';
-
-interface LayoutProps {
-    children: ReactNode 
-}
+import {LayoutProps} from '../interfaces'
 
 const Layout = ({children}: LayoutProps) => {
     return <div className='layout'>{children}</div>
